fix clean up race condition
After the client exits, xvfb-run kills the server, but doesn't unset XVFBPID.

clean_up()  is called as part of xvfb-run's EXIT trap handler, but if
the server has exited before clean_up can kill it, the kill will result
in an error.

This fix avoids this by ignoring any error from the attempt to kill the server.